Add IMAGE_ID template and update healthcheck probes
Add an IMAGE_ID environment variable, which we can then use later
in the Django app so we can see it on the site somewhere.

Also update the healthcheck URL to the proper one and modify the
probes.

Switch out the readinessProbe for a startupProbe instead so the
container doesn't go into a warning state when if first comes up
and then probe at 5s rather than the default of 10s

Change-Id: Iec3019693e383349fde0d1cfd98a24584a9bdc7c